feat: add server layout for auth and main amd protected pages
