+++ src/app/[locale]/(auth)/layout.tsx
@@ -1,14 +1,16 @@
+import { ReactNode } from 'react';
+
 import { getServerUser } from '@/server/get-server-user';
-import RestClientPageDefault from './[...params]/page';
 import { HeaderApp } from '@/widgets';
 
-export default async function RestClientPage() {
+export default async function AuthLayout({ children }: { children: ReactNode }) {
   const userServer = await getServerUser();
-  const user = userServer ? { name: userServer.name } : null;
+  const user = userServer ? { name: userServer.name ?? null } : null;
+
   return (
     <>
       <HeaderApp user={user} />
-      <RestClientPageDefault />;
+      {children}
     </>
   );
 }
