+++ src/app/[locale]/page.tsx
@@ -1,15 +1,19 @@
-'use client';
-
 import '@ant-design/v5-patch-for-react-19';
 import { Content } from 'antd/es/layout/layout';
-import { AboutCourseSection, HeroSection, TeamSection } from '@/widgets';
+import { AboutCourseSection, HeaderApp, HeroSection, TeamSection } from '@/widgets';
+import { getServerUser } from '@/server/get-server-user';
 
-export default function Home() {
+export default async function Home() {
+  const userServer = await getServerUser();
+  const user = userServer ? { name: userServer.name } : null;
   return (
-    <Content className="bg-animate">
-      <HeroSection />
-      <TeamSection />
-      <AboutCourseSection />
-    </Content>
+    <>
+      <HeaderApp user={user} />
+      <Content className="bg-animate">
+        <HeroSection user={user} />
+        <TeamSection />
+        <AboutCourseSection />
+      </Content>
+    </>
   );
 }
