+++ src/app/[locale]/variables/variables-client.tsx
@@ -1,0 +1,34 @@
+'use client';
+
+import '@ant-design/v5-patch-for-react-19';
+import { Flex } from 'antd';
+import { Content } from 'antd/es/layout/layout';
+import Title from 'antd/es/typography/Title';
+import { CSSProperties, lazy, Suspense } from 'react';
+import { useTranslations } from 'next-intl';
+import { LoadingSpinner } from '@/shared/UI';
+
+const EditableTable = lazy(() => import('@/features/editable-table'));
+
+const contentStyles: CSSProperties = {
+  maxWidth: '1440px',
+  margin: '0 auto',
+  padding: '20px 16px',
+  height: '100%',
+  textAlign: 'center',
+};
+
+export function VariablesView() {
+  const t = useTranslations('Variables');
+
+  return (
+    <Suspense fallback={<LoadingSpinner tip={t('loading')} />}>
+      <Content>
+        <Flex vertical gap={20} style={contentStyles}>
+          <Title>{t('title')}</Title>
+          <EditableTable />
+        </Flex>
+      </Content>
+    </Suspense>
+  );
+}
